Fix acceptance-window truncation in KeeperStoryPipeline getter
KeeperStoryPipeline::getAcceptanceWindow() returned uint16_t while the
acceptanceWindow member is uint64_t stored in nanoseconds (ctor does
acceptanceWindow *= 1000000000). Even a 1s window (1e9 ns) was truncated to
its low 16 bits, so the getter returned <=65535 ns (~tens of us) of garbage
instead of the real window (e.g. 15s -> 54784 ns, 180s -> 2048 ns).

The getter feeds KeeperDataStore's pipeline retirement: exit_time = now +
getAcceptanceWindow(). Truncation made exit_time ~= now + ~us, so a released
story's pipeline was retired at the next ~1s tick instead of being kept for the
acceptance window. That nullified the grace period meant to absorb in-flight /
late record_event RPCs after release; such events would arrive after the
ingestion handle was removed, get orphaned, and never be sealed.

The internal seal logic (extractDecayedStoryChunks) uses the uint64_t member
directly, so chunk sealing and tail reads were unaffected. The sibling
chrono-common/StoryPipeline.h getter is already correctly uint64_t.

Verified: exit_time - release_time is now exactly the configured window
(15.000 s), and the released pipeline is retired ~15 s after release instead of
~1 s.

+++ src/chrono-keeper/include/KeeperStoryPipeline.h
@@ -57,7 +57,7 @@ public:
 
     StoryName const& getStoryName() const { return storyName; }
 
-    uint16_t getAcceptanceWindow() const { return acceptanceWindow; }
+    uint64_t getAcceptanceWindow() const { return acceptanceWindow; }
 
     uint64_t TimelineStart() const { return (*storyTimelineMap.begin()).first; } // storyTimelineMap is never left empty
